Further Phrase Commands (#10)
* Apply Pin Widgets on Selection

Applies the Pin Widgets to the Owner Graph Node, on Selection of the Node.

Changed Parse Record to use a MultiMap, to allow for easier Phrase Mapping using duplicate Phrases.

* Added Initial Localised View Locomotion

Added an Initial Implementation of Localized View Locomotion. With their bound commands.

Current Bound Commands Being:
View > Move
View > Zoom
View > Focus - Uses Local Index

* Added Delegate Support On Derrived Input Nodes

Added Support For the New Input Delegate Event, in the constructor for derived input classes.

* Update PhraseEventNode.cpp

Missed Delegate Name Change

* Fixed Debug Errors

Fixed various errors that occurred during debugging, by further implementing data checks.

And Opting towards using a Data copy for certain propagation events, due to the possibility of needing to restart.

* Added Non-Numeric String Conversion

Added String Conversion for Numeric Values in a Word Format.
The conversion happens on all inputs that are detected as having numeric conversions, and are converted inline.

Example:
"ZERO" -> "0"

* Updated Meta Multipart Send

Updated the Multipart Send with Meta to a better method for list concatenation, as this one iteratively adds the values.

* Revert "Updated Meta Multipart Send"

This reverts commit 104f6fba7bc3b7804599cd83bc767f469c135745.

* Resolved Phrase Tree Parsing Issue.

Resolved an Issue when Phrase Tree Parsing, that was introduced in solving a current edge case (Will be fixed in later iteration).

* Fixed Command Events

Fixed the Command Events, due to the mistyping of the Lambda Bind.
Additionally, changed the events to work with the schema, so the performed mappings are performable by type.