Conflict On Commands due to Levenshtein distance
Current Conflit on Current Commands Found, due to the Levenshtein Distance allowing for acceptance of the first found node. Where Nodes Mapped to "Move" and "Remove" get treated the same.

Issue will be opened for a fix by usage of full child loop, and then using the lowest distancing child for further propagation.