Moved Command Implementations
Moves the Compile and Movement commands to a new "NODE" sister command "GRAPH" due to "NODE" being too focused on the contents of the graph, when these commands are more general graph functions.

Also moved the Close Command For Base Context Objects, to use a default solution appended to the PhraseContextNode.